requirements.txt
django
djangorestframework
python-dotenv
whitenoise
Pillow

psycopg2-binary
dj-database-url
celery
apscheduler
openai
jinja2
pytest
pytest-django
pytest-cov
anyio
requests
django-cors-headers
django-rest-passwordreset